# api/openapi/swagger.yaml
basePath: /
definitions:
...
  domain.EnvVar:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
...
  ? github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_workloadconfig_domain_WorkloadConfig
  : properties:
      data:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_workloadconfig_domain.WorkloadConfig'
    type: object
...
  ? github_com_bsonger_devflow-service_internal_platform_httpx.ListResponse-github_com_bsonger_devflow-service_internal_workloadconfig_domain_WorkloadConfig
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_workloadconfig_domain.WorkloadConfig'
        type: array
      pagination:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.Pagination'
    type: object
  github_com_bsonger_devflow-service_internal_platform_httpx.Pagination:
    properties:
      enabled:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_bsonger_devflow-service_internal_workloadconfig_domain.LabelItem:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  github_com_bsonger_devflow-service_internal_workloadconfig_domain.WorkloadConfig:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      env:
        items:
          $ref: '#/definitions/domain.EnvVar'
        type: array
      id:
        type: string
      labels:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_workloadconfig_domain.LabelItem'
        type: array
      name:
        type: string
      probes:
        additionalProperties: {}
        type: object
      replicas:
        type: integer
      resources:
        additionalProperties: {}
        type: object
      service_account_name:
        type: string
      strategy:
        type: string
      updated_at:
        type: string
      workload_type:
        type: string
    type: object
  github_com_bsonger_devflow-service_internal_workloadconfig_domain.WorkloadConfigInput:
    properties:
      application_id:
        type: string
      description:
        type: string
      env:
        items:
          $ref: '#/definitions/domain.EnvVar'
        type: array
      labels:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_workloadconfig_domain.LabelItem'
        type: array
      name:
        type: string
      probes:
        additionalProperties: {}
        type: object
      replicas:
        type: integer
      resources:
        additionalProperties: {}
        type: object
      service_account_name:
        type: string
      strategy:
        type: string
      workload_type:
        type: string
    type: object
...
info:
  contact: {}
  description: DevFlow meta-service HTTP API.
  title: DevFlow Meta Service API
  version: "1.0"
paths:
...
  /api/v1/workload-configs:
    get:
      parameters:
      - description: Application ID
        in: query
        name: application_id
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ListResponse-github_com_bsonger_devflow-service_internal_workloadconfig_domain_WorkloadConfig'
      summary: List workload configs
      tags:
      - WorkloadConfig
    post:
      consumes:
      - application/json
      parameters:
      - description: WorkloadConfig data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_workloadconfig_domain.WorkloadConfigInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_workloadconfig_domain_WorkloadConfig'
      summary: Create workload config
      tags:
      - WorkloadConfig
  /api/v1/workload-configs/{id}:
    delete:
      parameters:
      - description: WorkloadConfig ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete workload config
      tags:
      - WorkloadConfig
    get:
      parameters:
      - description: WorkloadConfig ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_workloadconfig_domain_WorkloadConfig'
      summary: Get workload config
      tags:
      - WorkloadConfig
    put:
      consumes:
      - application/json
      parameters:
      - description: WorkloadConfig ID
        in: path
        name: id
        required: true
        type: string
      - description: WorkloadConfig data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_workloadconfig_domain.WorkloadConfigInput'
      responses:
        "204":
          description: No Content
      summary: Update workload config
      tags:
      - WorkloadConfig
swagger: "2.0"